Riallocamento del referral, merge della CustomerActivityCard into CustomerCard
