+++ src/main/java/it/unicam/cs/ids/lp/client/card/CardProgram.java
@@ -1,0 +1,8 @@
+package it.unicam.cs.ids.lp.client.card;
+
+public enum CardProgram {
+    POINTS,
+    LEVELS,
+    MEMBERSHIP,
+    CASHBACK
+}
