+++ src/main/java/it/unicam/cs/ids/lp/client/card/CustomerCardMapper.java
@@ -1,5 +1,6 @@
 package it.unicam.cs.ids.lp.client.card;
 
+import it.unicam.cs.ids.lp.activity.card.CardRepository;
 import it.unicam.cs.ids.lp.client.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,10 +13,19 @@ public class CustomerCardMapper implements Function<CustomerCardRequest, Custome
     @Autowired
     private CustomerRepository customerRepository;
 
+    @Autowired
+    private CardRepository cardRepository;
+
     @Override
     public CustomerCard apply(CustomerCardRequest customerCardRequest) {
         CustomerCard customerCard = new CustomerCard();
         customerCard.setCustomer(customerRepository.findById(customerCardRequest.customerId()).orElseThrow());
+        customerCard.setCard(cardRepository.findById(customerCardRequest.cardId()).orElseThrow());
+        customerCard.setProgram(customerCardRequest.program());
+        customerCard.setFamily(customerCardRequest.family());
+        customerCard.setReferred(customerCardRequest.referredCode());
+
         return customerCard;
     }
 }
+
