+++ src/main/java/it/unicam/cs/ids/lp/client/card/CustomerCardRequest.java
@@ -1,4 +1,9 @@
 package it.unicam.cs.ids.lp.client.card;
 
-record CustomerCardRequest(long customerId) {
+public record CustomerCardRequest(long customerId,
+                                  long cardId,
+                                  CardProgram program,
+                                  boolean family,
+                                  String referredCode) {
 }
+
