--- src/main/java/it/unicam/cs/ids/lp/client/card/activitycard/CACardProgram.java
@@ -1,8 +1,0 @@
-package it.unicam.cs.ids.lp.client.card.activitycard;
-
-public enum CACardProgram {
-    POINTS,
-    LEVELS,
-    MEMBERSHIP,
-    CASHBACK
-}
